requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bezier==2021.2.12
blessings==1.7
cachetools==5.0.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coloredlogs==15.0.1
cssselect2==0.5.0
cssutils==2.4.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diffvg==0.0.1
einops==0.4.1
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
greenlet==1.1.2
grpcio==1.43.0
humanfriendly==10.0
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.1
ipykernel==6.9.1
ipython==8.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
kornia==0.1.4
llvmlite==0.38.0
lpips==0.1.4
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mpmath==1.2.1
multidict==6.0.2
nest-asyncio==1.5.4
networkx==2.6.3
numba==0.55.1
numpy==1.21.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
objprint==0.2.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prettytable==3.3.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytorch-ranger==0.1.1
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.2.0
svgpathtools==1.4.4
svgwrite==1.4.1
sympy==1.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
threadpoolctl==3.1.0
tifffile==2022.2.9
tinycss2==1.1.1
torch==1.9.0+cu111
torch-optimizer==0.3.0
torch-tools==0.1.5
torchfile==0.1.0
torchinfo==1.7.1
torchmetrics==0.7.2
torchsummary==1.5.1
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
